Some refactoring menuBundle - Make it extendable - thank you Kristof and Wim
improvements to extendability

discovered the wonders of bundle DI

removed getClass methods

small fixes and styling

reverted changes

fixes after feedback wim
