+++ src/Kunstmaan/MenuBundle/Controller/MenuAdminListController.php
@@ -26,7 +26,8 @@ class MenuAdminListController extends AdminListController
     public function getAdminListConfigurator(Request $request)
     {
         if (!isset($this->configurator)) {
-            $this->configurator = new MenuAdminListConfigurator(
+	    $configuratorClass = $this->getParameter('kunstmaan_menu.adminlist.menu_configurator.class');
+	    $this->configurator = new $configuratorClass(
                 $this->getEntityManager()
             );
