+++ src/Kunstmaan/MenuBundle/Repository/MenuItemRepositoryInterface.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: bart
+ * Date: 09/10/15
+ * Time: 14:38
+ */
+
+namespace Kunstmaan\MenuBundle\Repository;
+
+
+interface MenuItemRepositoryInterface
+{
+    /**
+     * @param string $menuName
+     * @param string $locale
+     * @return array
+     */
+    public function getMenuItemsForLanguage($menuName, $locale);
+
+}
